# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get the authenticated user's profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not yet provisioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update the authenticated user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    post:
      operationId: provisionUser
      tags: [users]
      summary: Provision the authenticated user's local account
      description: Creates the local user record bound to the authenticated identity, with the chosen role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: Provisioned user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Provider provisioning requires commission agreement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves object entities uploaded via presigned URLs. These can optionally
        be protected with authentication or ACL checks based on the use case.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir (e.g. `uploads/some-uuid`).
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, authUserId, name, role, createdAt, updatedAt]
      properties:
        id:
          type: integer
        authUserId:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, provider, admin]
        avatarUrl:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        commissionAgreedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          enum: [customer, provider]
        email:
          type: string
        phone:
          type: string
        commissionAgreed:
          type: boolean
          description: Required true when role is "provider"; records consent to the platform commission.
        language:
          type: string
          enum: [ar, en, he]
          description: Preferred UI language; used to localize server-sent notifications.
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
        avatarUrl:
          type: string
        address:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
...